% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiProbitPartialR.R
\docType{package}
\name{BiProbitPartial-package}
\alias{BiProbitPartial-package}
\alias{_PACKAGE}
\title{BiProbitPartial: Bivariate Probit with Partial Observability}
\description{
A suite of functions to estimate, summarize
  and perform predictions with the bivariate probit subject to partial observability.
  The frequentist and Bayesian probabilistic philosophies are both supported. The 
  frequentist method is estimated with maximum likelihood and the Bayesian method is 
  estimated with a Markov Chain Monte Carlo (MCMC) algorithm developed by 
  Rajbanhdari, A (2014) <doi:10.1002/9781118771051.ch13>.
}
