% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodfunctions.R
\name{grad1}
\alias{grad1}
\title{Gradient of bivariate probit with partial observability}
\usage{
grad1(theta, X1, X2, Z, rho = 0, p = NULL, summed = T, fixrho = F)
}
\arguments{
\item{theta}{numeric vector of dimension equal to that of the free parameter space}

\item{X1}{numeric matrix of covariates for the first equation}

\item{X2}{numeric matrix of covariates for the second equation}

\item{Z}{numeric matrix or column vecotr of response observations}

\item{rho}{numeric value for rho if fixed}

\item{p}{numeric precomputed probabilities of Pr(Y1=1,Y2=1)}

\item{summed}{logical if the gradient observations should be summed}

\item{fixrho}{logical if rho should be fixed}
}
\value{
if summed is TRUE then the function returns the numeric column sum of the gradient matrix, else it returns a numeric vector with each entry a value of the gradient vector
}
\description{
Gradient of bivariate probit with partial observability
}
