% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiProbitPartialR.R
\docType{data}
\name{SimDat}
\alias{SimDat}
\title{This is data to be included in my package}
\description{
Simulated data of 10,000 observations from a multivariate
normal distribution. The true coefficients for equation 1 are 0 and 1. The
true coefficients for equation 2 are 0 and -1. The true \eqn{\rho} is 0.
}
\author{
Michael Guggisberg \email{mguggisb@ida.org}
}
\keyword{data}
